The code now runs. The main issue solved was with the computation of KL divergence, adding an eps to prevent the generation of NaNs, which then overflowed to all parameters during backpropagation

+++ dataset/.ipynb_checkpoints/dataset-checkpoint.py
@@ -152,7 +152,7 @@ class Dataset:
                 names = np.unique(values)
                 self.num_covariates.append(len(names))
 
-                names_idx = torch.arange(len(names), dtype=torch.float32).unsqueeze(-1)
+                names_idx = torch.arange(len(names)).unsqueeze(-1)
                 self.covars_dict[cov] = dict(
                     zip(list(names), names_idx)
                 )
